1) Base and Simple data providers. 2) UT for Simple data provider
