--- modules/bb_dir/bb_file.py
@@ -1,7 +1,0 @@
-class BB1 ():
-    def b(self):
-        return "b-1"
-
-class BB2 ():
-    def b(self):
-        return "b-2"
