--- source/bars/__init__.py
@@ -1,3 +1,0 @@
-__all__ = [
-    "bars.bar"
-    ]
