--- source/bars/bar.py
@@ -1,4 +1,0 @@
-class Bar():
-
-    def bar(self):
-        return "bar"
