+++ source/data_providers/__init__.py
@@ -1,3 +1,4 @@
 __all__ = [
-    "source.data_providers.data_provider_base"
+    "data_provider_base",
+    "data_provider_simple"
     ]
