+++ source/data_providers/data_provider_base.py
@@ -1,10 +1,6 @@
 class DataProviderBase:
-    result = []
-    def get_data(self):
-        return result
+    result = None
+    def get_data(self) -> list:
+        return self.result 
 
-
-class DataProviderSimple ( DataProviderBase ):
-
-    def __init__( self ):
-        self.result = [1, 2, 3]
+
