+++ source/data_providers/data_provider_simple.py
@@ -1,0 +1,6 @@
+import data_provider_base as dpb
+
+class DataProviderSimple ( dpb.DataProviderBase ):
+
+    def __init__( self ):
+        self.result = [0, 1, 2, 3]
