--- source/foo.py
@@ -1,4 +1,0 @@
-class Foo():
-
-    def foo(self):
-        return "foo"
