+++ unit_tests/data_provider_simple/__init__.py
@@ -1,3 +1,3 @@
 __all__ = [
-    "bb_file"
+    "env"
     ]
