--- unit_tests/env.py
@@ -1,15 +1,0 @@
-import sys
-import os
-
-
-print(sys.path)
-
-# append module root directory to sys.path
-sys.path.append(
-    os.path.dirname(
-        os.path.dirname(
-            os.path.abspath(__file__)
-        )
-    )
-)
-print(sys.path)
